입찰 삭제시 권한 검사

+++ src/main/java/com/api/farmingsoon/domain/bid/service/BidService.java
@@ -43,8 +43,8 @@ public class BidService {
 
     @Transactional
     public void delete(Long bidId) {
-        Member member = authenticationUtils.getAuthenticationMember();
         Bid bid = bidRepository.findById(bidId).orElseThrow(() -> new NotFoundException(ErrorCode.NOT_FOUND_BID));
+        AuthenticationUtils.checkDeletePermission(bid.getMember());
 
         bidRepository.delete(bid);
     }
